requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=1.4.0=pypi_0
aom=3.8.2=h59595ed_0
backcall=0.2.0=pyh9f0ad1d_0
blas=1.0=mkl
brotli=1.1.0=hd590300_1
brotli-bin=1.1.0=hd590300_1
brotli-python=1.0.9=py37hd23a5d3_7
bzip2=1.0.8=hd590300_5
ca-certificates=2024.2.2=hbcca054_0
cachetools=5.3.1=pypi_0
cairo=1.18.0=h3faef2a_0
certifi=2023.5.7=pypi_0
cffi=1.15.1=py37h43b0acd_1
charset-normalizer=3.2.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
conda=4.12.0=py37h89c1867_0
conda-package-handling=1.9.0=py37h540881e_0
cryptography=38.0.2=py37h5994e8b_1
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pyhd8ed1ab_0
dav1d=1.2.1=hd590300_0
debugpy=1.7.0=pypi_0
decorator=5.1.1=pyhd8ed1ab_0
deepsnap=0.2.1=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
expat=2.6.2=h59595ed_0
ffmpeg=6.1.1=gpl_h38e077a_106
fftw=3.3.10=nompi_hc118613_108
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_1
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h267a509_2
fribidi=1.0.10=h36c2ea0_0
future=0.18.3=pypi_0
gettext=0.21.1=h27087fc_0
giflib=5.2.1=h0b41bf4_3
gmp=6.3.0=h59595ed_1
gnutls=3.7.9=hb077bed_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.56.0=pypi_0
harfbuzz=8.3.0=h3d44ed6_0
icu=73.2=h59595ed_0
idna=3.4=pypi_0
importlib-metadata=6.7.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.16.2=pypi_0
ipython=7.33.0=py37h89c1867_0
ipython_genutils=0.2.0=py_1
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.3.1=pypi_0
jpeg=9b=h024ee3a_2
jupyter-client=7.4.9=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter_core=4.11.1=py37h89c1867_0
kiwisolver=1.4.4=py37h6a678d5_0
lame=3.100=h166bdaf_1003
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h41732ed_0
libabseil=20240116.1=cxx17_h59595ed_2
libass=0.17.1=h8fe9dca_1
libbrotlicommon=1.1.0=hd590300_1
libbrotlidec=1.1.0=hd590300_1
libbrotlienc=1.1.0=hd590300_1
libdrm=2.4.120=hd590300_0
libexpat=2.6.2=h59595ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_5
libgfortran-ng=13.2.0=h69a702a_5
libgfortran5=13.2.0=ha4646dd_5
libglib=2.80.0=hf2295e7_1
libgomp=13.2.0=h807b86a_5
libhwloc=2.9.3=default_h554bfaf_1009
libiconv=1.17=hd590300_2
libidn2=2.3.7=hd590300_0
libnsl=2.0.1=hd590300_0
libopenvino=2024.0.0=h2e90f83_4
libopenvino-auto-batch-plugin=2024.0.0=hd5fc58b_4
libopenvino-auto-plugin=2024.0.0=hd5fc58b_4
libopenvino-hetero-plugin=2024.0.0=h3ecfda7_4
libopenvino-intel-cpu-plugin=2024.0.0=h2e90f83_4
libopenvino-intel-gpu-plugin=2024.0.0=h2e90f83_4
libopenvino-ir-frontend=2024.0.0=h3ecfda7_4
libopenvino-onnx-frontend=2024.0.0=h757c851_4
libopenvino-paddle-frontend=2024.0.0=h757c851_4
libopenvino-pytorch-frontend=2024.0.0=h59595ed_4
libopenvino-tensorflow-frontend=2024.0.0=hca94c1a_4
libopenvino-tensorflow-lite-frontend=2024.0.0=h59595ed_4
libopus=1.3.1=h7f98852_1
libpciaccess=0.18=hd590300_0
libpng=1.6.43=h2797004_0
libprotobuf=4.25.3=h08a7969_0
libsodium=1.0.18=h36c2ea0_1
libsolv=0.7.28=hfc55251_1
libsqlite=3.45.2=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_5
libtasn1=4.19.0=h166bdaf_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h7f98852_0
libuuid=2.38.1=h0b41bf4_0
libuv=1.48.0=hd590300_0
libva=2.21.0=hd590300_0
libvpx=1.14.0=h59595ed_0
libwebp=1.2.0=h89dd481_0
libxcb=1.15=h0b41bf4_0
libxml2=2.12.6=h232c23b_1
libzlib=1.2.13=hd590300_5
lz4-c=1.9.3=h9c3ff4c_1
mamba=0.1.2=py37h99015e2_0
markdown=3.4.3=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib-base=3.5.3=py37hf395dca_2
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4.20240210=h59595ed_0
nest-asyncio=1.6.0=pypi_0
nettle=3.9.1=h7ab15ed_0
networkx=2.6.3=pypi_0
ninja=1.11.1=h924138e_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
oauthlib=3.2.2=pypi_0
ocl-icd=2.3.2=hd590300_1
openh264=2.4.1=h59595ed_0
openssl=3.2.1=hd590300_1
p11-kit=0.24.1=hc5aa10d_0
packaging=23.1=pypi_0
pandas=1.3.5=py37he8f5f7f_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.6=pyhd8ed1ab_0
pcre2=10.43=hcad00b1_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.3.0=py37hace64e9_1
pip=24.0=pyhd8ed1ab_0
pixman=0.43.2=h59595ed_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=3.20.3=pypi_0
psutil=5.9.5=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pugixml=1.14=h59595ed_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycosat=0.6.4=py37h540881e_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pyopenssl=23.2.0=pyhd8ed1ab_1
pyparsing=3.1.0=pypi_0
pysocks=1.7.1=py37h89c1867_5
python=3.7.12=hf930737_100_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.7=4_cp37m
pytorch=1.9.0=py3.7_cuda10.2_cudnn7.6.5_0
pytz=2024.1=pyhd8ed1ab_0
pyzmq=24.0.1=py37h0c0c2a8_0
readline=8.2=h8228510_1
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
ruamel_yaml=0.15.80=py37h540881e_1007
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=py37h6c91a56_2
seaborn=0.12.2=hd8ed1ab_0
seaborn-base=0.12.2=pyhd8ed1ab_0
setuptools=59.5.0=pypi_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h9fff704_0
sqlite=3.45.2=h2c6b66d_0
statsmodels=0.13.5=py37h7deecbd_1
svt-av1=1.8.0=h59595ed_0
tb-nightly=2.12.0a20230113=pypi_0
tbb=2021.11.0=h00ab1b0_1
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
torch-cluster=1.5.9=pypi_0
torch-geometric=2.2.0=pypi_0
torch-scatter=2.0.7=pypi_0
torch-sparse=0.6.10=pypi_0
torch-spline-conv=1.2.1=pypi_0
torchaudio=0.9.0=py37
torchvision=0.10.0=py37_cu102
tornado=6.2=py37h540881e_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
urllib3=1.26.16=pypi_0
wcwidth=0.1.9=pyh9f0ad1d_0
werkzeug=2.2.3=pypi_0
wheel=0.43.0=pyhd8ed1ab_0
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xorg-fixesproto=5.0=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.7=h8ee46fc_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxrender=0.9.11=hd590300_0
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
zeromq=4.3.5=h59595ed_1
zipp=3.15.0=pypi_0
zlib=1.2.13=hd590300_5
zstd=1.4.9=ha95c52a_0
